putting dtype=float in everything just in case

+++ tedi/kernels.py
@@ -17,7 +17,7 @@ class kernel(object):
         """
             Puts all kernel arguments in an array pars.
         """
-        self.pars = np.array(args)
+        self.pars = np.array(args, dtype=float)
 
     def __call__(self, r):
         """
